requirements-render.txt
Flask==3.0.0
Werkzeug==3.0.1
striprtf==0.0.26
gunicorn==21.2.0
PyYAML==6.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click>=8.0
blinker>=1.1
itsdangerous>=2.0
